Refactor out to class checker
