+++ agent-backend/src/messaging/send_message_to_socket.py
@@ -2,5 +2,10 @@ from models.sockets import SocketMessage, SocketEvents
 from socketio import SimpleClient
 
 
-def send(client: SimpleClient, event: SocketEvents, message: SocketMessage):
-    client.emit(event.value, message.model_dump())
+def send(client: SimpleClient, 
+         event: SocketEvents, 
+         message: SocketMessage):
+    
+    # Check inputs
+    
+    client.emit(event=event.value, data=message.model_dump())
